# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Civic reporting and case-tracking API for Nagar Sarvadarshin
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  citizen: 'citizen',
  employee: 'employee',
  admin: 'admin',
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface RegisterInput {
  /**
     * @minLength 2
     * @maxLength 80
     */
  name: string;
  email: string;
  /**
     * @minLength 8
     * @maxLength 100
     */
  password: string;
}

export interface AuthSession {
  token: string;
  user: User;
}

export type CivicCasePriorityLevel = typeof CivicCasePriorityLevel[keyof typeof CivicCasePriorityLevel];


export const CivicCasePriorityLevel = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
  CRITICAL: 'CRITICAL',
} as const;

export interface CivicCase {
  id: string;
  caseNumber: string;
  title: string;
  description: string;
  category: string;
  /** @nullable */
  subcategory: string | null;
  latitude: number;
  longitude: number;
  address: string;
  department: string;
  priorityScore: number;
  priorityLevel: CivicCasePriorityLevel;
  status: string;
  reportCount: number;
  affectedPopulation: number;
  createdAt: string;
  updatedAt: string;
}

/**
 * @nullable
 */
export type ComplaintInputDuplicateAction = typeof ComplaintInputDuplicateAction[keyof typeof ComplaintInputDuplicateAction] | null;


export const ComplaintInputDuplicateAction = {
  link: 'link',
  new: 'new',
} as const;

export interface ComplaintInput {
  /**
     * @minLength 4
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 8
     * @maxLength 1000
     */
  description: string;
  /** @nullable */
  category?: string | null;
  /**
     * @minimum -90
     * @maximum 90
     */
  latitude: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  longitude: number;
  /**
     * @minimum 0
     * @nullable
     */
  accuracy?: number | null;
  /**
     * @minLength 1
     * @maxLength 240
     */
  address: string;
  /**
     * @maxLength 1500000
     * @nullable
     */
  imageData?: string | null;
  /**
     * @maxLength 1200
     * @nullable
     */
  voiceTranscript?: string | null;
  isProtected?: boolean;
  /** @nullable */
  duplicateAction?: ComplaintInputDuplicateAction;
  /** @nullable */
  linkCaseId?: string | null;
}

export type AIAnalysisPriorityLevel = typeof AIAnalysisPriorityLevel[keyof typeof AIAnalysisPriorityLevel];


export const AIAnalysisPriorityLevel = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
  CRITICAL: 'CRITICAL',
} as const;

export interface AIAnalysis {
  category: string;
  subcategory: string;
  severity: string;
  confidence: number;
  priorityScore: number;
  priorityLevel: AIAnalysisPriorityLevel;
  department: string;
  duplicateScore: number;
  reason: string;
}

export interface ComplaintSubmission {
  /** @nullable */
  complaintId: string | null;
  civicCase: CivicCase | null;
  candidateCase: CivicCase | null;
  awaitingDecision: boolean;
  isDuplicate: boolean;
  relatedReports: number;
  analysis: AIAnalysis;
}

export interface Notification {
  id: string;
  title: string;
  message: string;
  type: string;
  isRead: boolean;
  createdAt: string;
}