# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Civic reporting and case-tracking API for Nagar Sarvadarshin
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Demo account registration and sessions
  - name: civic
    description: Civic reports, cases, and map issues
  - name: notifications
    description: Citizen case updates
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with a demo or registered account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create a citizen account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "409":
          description: Email already registered
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the signed-in user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Authentication required
  /civic-cases:
    get:
      operationId: getCivicCases
      tags: [civic]
      summary: List visible civic cases
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cases visible to the signed-in user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CivicCase"
  /civic-cases/{id}:
    get:
      operationId: getCivicCase
      tags: [civic]
      summary: Get a civic case and its timeline
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Civic case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CivicCase"
        "404":
          description: Case not found
  /complaints:
    post:
      operationId: submitComplaint
      tags: [civic]
      summary: Analyze a report and create or link a civic case
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintInput"
      responses:
        "201":
          description: Report analyzed and attached to a case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintSubmission"
  /map/issues:
    get:
      operationId: getMapIssues
      tags: [civic]
      summary: List privacy-safe cases for the nearby map
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Map issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CivicCase"
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: List the signed-in user's notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [citizen, employee, admin] }
      required: [id, name, email, role]
    LoginInput:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
      required: [email, password]
    RegisterInput:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 80 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 100 }
      required: [name, email, password]
    AuthSession:
      type: object
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
      required: [token, user]
    CivicCase:
      type: object
      properties:
        id: { type: string }
        caseNumber: { type: string }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        subcategory: { type: ["string", "null"] }
        latitude: { type: number }
        longitude: { type: number }
        address: { type: string }
        department: { type: string }
        priorityScore: { type: integer }
        priorityLevel: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
        status: { type: string }
        reportCount: { type: integer }
        affectedPopulation: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - caseNumber
        - title
        - description
        - category
        - subcategory
        - latitude
        - longitude
        - address
        - department
        - priorityScore
        - priorityLevel
        - status
        - reportCount
        - affectedPopulation
        - createdAt
        - updatedAt
    ComplaintInput:
      type: object
      properties:
        title: { type: string, minLength: 4, maxLength: 120 }
        description: { type: string, minLength: 8, maxLength: 1000 }
        category: { type: ["string", "null"] }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        accuracy: { type: ["number", "null"], minimum: 0 }
        address: { type: string, minLength: 1, maxLength: 240 }
        imageData: { type: ["string", "null"], maxLength: 1500000 }
        voiceTranscript: { type: ["string", "null"], maxLength: 1200 }
        isProtected: { type: boolean }
        duplicateAction: { type: ["string", "null"], enum: [link, new] }
        linkCaseId: { type: ["string", "null"] }
      required: [title, description, latitude, longitude, address]
    AIAnalysis:
      type: object
      properties:
        category: { type: string }
        subcategory: { type: string }
        severity: { type: string }
        confidence: { type: integer }
        priorityScore: { type: integer }
        priorityLevel: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
        department: { type: string }
        duplicateScore: { type: integer }
        reason: { type: string }
      required:
        - category
        - subcategory
        - severity
        - confidence
        - priorityScore
        - priorityLevel
        - department
        - duplicateScore
        - reason
    ComplaintSubmission:
      type: object
      properties:
        complaintId: { type: ["string", "null"] }
        civicCase: { anyOf: [{ $ref: "#/components/schemas/CivicCase" }, { type: "null" }] }
        candidateCase: { anyOf: [{ $ref: "#/components/schemas/CivicCase" }, { type: "null" }] }
        awaitingDecision: { type: boolean }
        isDuplicate: { type: boolean }
        relatedReports: { type: integer }
        analysis: { $ref: "#/components/schemas/AIAnalysis" }
      required: [complaintId, civicCase, candidateCase, awaitingDecision, isDuplicate, relatedReports, analysis]
    Notification:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        message: { type: string }
        type: { type: string }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, title, message, type, isRead, createdAt]