# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Civic reporting and case-tracking API for Nagar Sarvadarshin
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthSession {
  token: string;
  user: User;
}