# lib/api-zod/src/generated/types/civicCase.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Civic reporting and case-tracking API for Nagar Sarvadarshin
 * OpenAPI spec version: 0.1.0
 */
import type { CivicCasePriorityLevel } from './civicCasePriorityLevel';

export interface CivicCase {
  id: string;
  caseNumber: string;
  title: string;
  description: string;
  category: string;
  /** @nullable */
  subcategory: string | null;
  latitude: number;
  longitude: number;
  address: string;
  department: string;
  priorityScore: number;
  priorityLevel: CivicCasePriorityLevel;
  status: string;
  reportCount: number;
  affectedPopulation: number;
  createdAt: Date;
  updatedAt: Date;
}