# lib/api-zod/src/generated/types/complaintInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Civic reporting and case-tracking API for Nagar Sarvadarshin
 * OpenAPI spec version: 0.1.0
 */
import type { ComplaintInputDuplicateAction } from './complaintInputDuplicateAction';

export interface ComplaintInput {
  /**
     * @minLength 4
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 8
     * @maxLength 1000
     */
  description: string;
  /** @nullable */
  category?: string | null;
  /**
     * @minimum -90
     * @maximum 90
     */
  latitude: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  longitude: number;
  /**
     * @minimum 0
     * @nullable
     */
  accuracy?: number | null;
  /**
     * @minLength 1
     * @maxLength 240
     */
  address: string;
  /**
     * @maxLength 1500000
     * @nullable
     */
  imageData?: string | null;
  /**
     * @maxLength 1200
     * @nullable
     */
  voiceTranscript?: string | null;
  isProtected?: boolean;
  /** @nullable */
  duplicateAction?: ComplaintInputDuplicateAction;
  /** @nullable */
  linkCaseId?: string | null;
}