# lib/api-zod/src/generated/types/complaintSubmission.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Civic reporting and case-tracking API for Nagar Sarvadarshin
 * OpenAPI spec version: 0.1.0
 */
import type { AIAnalysis } from './aIAnalysis';
import type { CivicCase } from './civicCase';

export interface ComplaintSubmission {
  /** @nullable */
  complaintId: string | null;
  civicCase: CivicCase | null;
  candidateCase: CivicCase | null;
  awaitingDecision: boolean;
  isDuplicate: boolean;
  relatedReports: number;
  analysis: AIAnalysis;
}