# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Civic reporting and case-tracking API for Nagar Sarvadarshin
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}