# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Civic reporting and case-tracking API for Nagar Sarvadarshin
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /**
     * @minLength 2
     * @maxLength 80
     */
  name: string;
  email: string;
  /**
     * @minLength 8
     * @maxLength 100
     */
  password: string;
}