# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Civic reporting and case-tracking API for Nagar Sarvadarshin
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
}